fix bug

+++ src/storage/src/redis_strings.cc
@@ -1767,7 +1767,7 @@ rocksdb::Status Redis::PKPatternMatchDel(const std::string& pattern, int32_t* re
   return s;
 }
 
-rocksdb::Status Redis::PKPatternMatchDelWithRemoveKeys(const std::string& pattern, int64_t* ret, std::vector<std::string>* remove_keys, const uint64_t& max_count) {
+rocksdb::Status Redis::PKPatternMatchDelWithRemoveKeys(const std::string& pattern, int64_t* ret, std::vector<std::string>* remove_keys, const int64_t& max_count) {
   rocksdb::ReadOptions iterator_options;
   const rocksdb::Snapshot* snapshot;
   ScopeSnapshot ss(db_, &snapshot);
@@ -1776,12 +1776,12 @@ rocksdb::Status Redis::PKPatternMatchDelWithRemoveKeys(const std::string& patter
 
   std::string key;
   std::string meta_value;
-  int32_t total_delete = 0;
+  int64_t total_delete = 0;
   rocksdb::Status s;
   rocksdb::WriteBatch batch;
   rocksdb::Iterator* iter = db_->NewIterator(iterator_options, handles_[kMetaCF]);
   iter->SeekToFirst();
-  while (iter->Valid() && static_cast<size_t>(batch.Count()) < max_count) {
+  while (iter->Valid() && static_cast<int64_t>(batch.Count()) < max_count) {
     auto meta_type = static_cast<enum DataType>(static_cast<uint8_t>(iter->value()[0]));
     ParsedBaseMetaKey parsed_meta_key(iter->key().ToString());
     key = iter->key().ToString();
@@ -1827,7 +1827,7 @@ rocksdb::Status Redis::PKPatternMatchDelWithRemoveKeys(const std::string& patter
   if (batch.Count() != 0U) {
     s = db_->Write(default_write_options_, &batch);
     if (s.ok()) {
-      total_delete += static_cast<int32_t>(batch.Count());
+      total_delete += static_cast<int64_t>(batch.Count());
       batch.Clear();
     } else {
       remove_keys->erase(remove_keys->end() - batch.Count(), remove_keys->end());
